Cleaned up the title replacement
Rewrote the main function and added new functions. Now, the file is first split at the xml-level (the 'front' node is removed completely). The result of that is fed into the xslt-applying function. Then after, the title-generating function creates a title according to BMGN's format based on the original xml. The title and the rest are then merged.